Removed mess and added gamecontroller which can be used for othello
