+++ src/main/java/isy/team4/projectisy/controller/OthelloController.java
@@ -1,3 +1,4 @@
 package isy.team4.projectisy.controller;
-public abstract class Controller {
+
+public class OthelloController {
 }
