+++ src/main/java/isy/team4/projectisy/controller/ReversiController.java
@@ -1,4 +1,4 @@
 package isy.team4.projectisy.controller;
 
-public class ReversiController extends Controller {
+public class ReversiController extends GameController {
 }
