+++ src/main/resources/isy/team4/projectisy/othello-view.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<VBox alignment="CENTER" prefHeight="429.0" prefWidth="447.0" spacing="20.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="isy.team4.projectisy.controller.TicTacToeController">
+<VBox alignment="CENTER" prefHeight="429.0" prefWidth="447.0" spacing="20.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="isy.team4.projectisy.controller.OthelloController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
     </padding>
@@ -81,4 +81,5 @@
     </GridPane>
     <Button cancelButton="true" onAction="#navigateToHome" text="Homepagina" />
    <Text fx:id="gameinfo" strokeType="OUTSIDE" strokeWidth="0.0" />
+    <Label fx:id="test" />
 </VBox>
